\name{antiviral_drug}
\alias{antiviral_drug}

\title{
List of antiviral drugs with SMILES strings
}

\description{
81 antiviral drugs with SMILES strings
}

\usage{
antiviral_drug
}

\value{
SMILES string
}

\source{
Huang, K., Fu, T., Glass, L. M., Zitnik, M., Xiao, C., & Sun, J. (2020). DeepPurpose: A Deep Learning Library for Drug-Target Interaction Prediction. Bioinformatics.
}

\author{
Dongmin Jung
}

