\name{example_bioassay}
\alias{example_bioassay}

\title{
Example Data for PubChem AID1706 bioassay
}

\description{
This is a compound-protein interaction data set retrieved from PubChem AID1706 bioassay. The data is balanced and a randomly selected subset of a dataset of size 5000. The label is 1 if the score is greater than or equal to 15, otherwise it is 0.
}

\usage{
example_bioassay
}

\value{
compound-protein interaction data
}

\source{
Huang, K., Fu, T., Glass, L. M., Zitnik, M., Xiao, C., & Sun, J. (2020). DeepPurpose: A Deep Learning Library for Drug-Target Interaction Prediction. Bioinformatics.
}

\author{
Dongmin Jung
}

