\name{example_cci}
\alias{example_cci}

\title{
Example Data for Chemical-Chemical Interactions
}

\description{
The data is a randomly selected subset with size 1000 for chemical-chemical interactions. The two SMILES strings are for compound pairs and the label is for their interactions.
}

\usage{
example_cci
}

\value{
chemical-chemical interaction data
}

\source{
Huang, K., Xiao, C., Hoang, T., Glass, L., & Sun, J. (2020). CASTER: Predicting drug interactions with chemical substructure representation. AAAI.
}

\author{
Dongmin Jung
}

