\name{example_cpi}
\alias{example_cpi}

\title{
Example Data for Compound-Protein Interactions
}

\description{
The data consist of compound-protein pairs and their interactions of human. The SMILES and amino acid sequences are used for compounds and proteins, respectively. The binary outcome label is whether or not they interact each other.
}

\usage{
example_cpi
}

\value{
compound-protein interaction data
}

\source{
Tsubaki, M., Tomii, K., & Sese, J. (2019). Compound-protein interaction prediction with end-to-end learning of neural networks for graphs and sequences. Bioinformatics, 35(2), 309-318.
}

\author{
Dongmin Jung
}

