\name{example_pd}
\alias{example_pd}

\title{
Example Data for Primer-Dimer
}

\description{
This is a primer-primer interaction data set with size 319. The two sequences are for primer pairs and the label is for their interactions.
}

\usage{
example_pd
}

\value{
primer sequences and dimer formation data
}

\source{
Johnston, A. D., Lu, J., Ru, K. L., Korbie, D., & Trau, M. (2019). PrimerROC: accurate condition-independent dimer prediction using ROC analysis. Scientific reports.
}

\author{
Dongmin Jung
}

