\name{example_ppi}
\alias{example_ppi}

\title{
Example Data for Protein-Protein Interactions
}

\description{
The data is a randomly selected subset with size 5000 for protein-protein interactions of yeast. The two amino acid sequences are for protein pairs and the label is for their interactions.
}

\usage{
example_ppi
}

\value{
protein-protein interaction data
}

\source{
Chen, M., et al. (2019). Multifaceted protein-protein interaction prediction based on siamese residual rcnn. Bioinformatics, 35(14), i305-i314.
}

\author{
Dongmin Jung
}

