\name{example_prot}
\alias{example_prot}

\title{
Example Data for Proteins
}

\description{
This is a protein data set retrieved from Research Collaboratory for Structural Bioinformatics (RCSB) Protein Data Bank (PDB).The data consist of amino acid sequences with three classes. The row name labels each row with the PDB identification code.
}

\usage{
example_prot
}

\value{
protein data
}

\source{
Research Collaboratory for Structural Bioinformatics (RCSB) Protein Data Bank (PDB) and https://www.kaggle.com/shahir/protein-data-set
}

\author{
Dongmin Jung
}

