\name{SARS_CoV2_3CL_Protease}
\alias{SARS_CoV2_3CL_Protease}

\title{
Amino Acid Sequence for the SARS coronavirus 3C-like Protease
}

\description{
306 amino acid residues of the SARS coronavirus 3C-like Protease
}

\usage{
SARS_CoV2_3CL_Protease
}

\value{
amino acid sequence
}

\source{
Huang, K., Fu, T., Glass, L. M., Zitnik, M., Xiao, C., & Sun, J. (2020). DeepPurpose: A Deep Learning Library for Drug-Target Interaction Prediction. Bioinformatics.
}

\author{
Dongmin Jung
}

